+++ config/jest.config.js
@@ -31,12 +31,7 @@ module.exports = (disabledMocks = []) => ({
             : { serialport: `${mockDir}/emptyMock.ts` }),
     },
     transform: {
-        '^.+\\.[jt]sx?$': [
-            'babel-jest',
-            {
-                configFile: `${__dirname}/babel.config.js`,
-            },
-        ],
+        '^.+\\.[jt]sx?$': '@swc/jest',
     },
     transformIgnorePatterns: ['node_modules/(?!(pc-nrfconnect-shared)/)'],
     setupFilesAfterEnv: [`${__dirname}/../test/setupTests.ts`],