+++ mocks/electronMock.ts
@@ -6,7 +6,6 @@
 
 // Allows jest to test files that import the electron package.
 
-export const require = jest.fn();
 export const match = jest.fn();
 export const app = jest.fn();
 export const dialog = jest.fn();